requirements.txt
asgiref==3.9.1
click==8.2.1
Django==5.2.4
django-cors-headers==4.7.0
django-redis==6.0.0
django-rq==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-django==4.11.1
python-decouple==3.8
python-dotenv==1.1.1
redis==6.2.0
rq==2.4.1
sqlparse==0.5.3
whitenoise==6.9.0
